requirements.txt
aiohttp==3.9.3
aiohttp-retry==2.8.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
attrs==23.2.0
blinker==1.7.0
certifi==2024.2.2
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
distlib==0.3.8
distro==1.9.0
dnspython==2.6.1
filelock==3.13.1
Flask[async]
frozenlist==1.4.1
googletrans
gunicorn
h11
h2
hpack
hstspreload
httpcore
httpx
hyperframe
idna
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
multidict==6.0.5
mysql-connector-python
opentelemetry-ap
opentelemetry-sdk
opentelemetry-exporter-gcp-trace
platformdirs==4.1.0
pydantic==2.6.4
pydantic_core==2.16.3
PyJWT==2.8.0
pymongo==4.6.3
requests==2.31.0
rfc3986==1.5.0
setuptools==69.0.3
sniffio==1.3.1
stripe==8.10.0
tqdm==4.66.2
twilio==9.0.2
typing_extensions==4.11.0
urllib3==2.2.1
virtualenv==20.25.0
Werkzeug==3.0.1
wheel==0.42.0
yarl==1.9.4
